Test: HouseGeo에서 Mock 테스트 분리
