+++ src/test/java/com/ssafy/tott/global/config/MockitoTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.context.ActiveProfiles;
 
-@ActiveProfiles(profiles = {"test"})
+@ActiveProfiles("test")
 @ExtendWith(MockitoExtension.class)
 public abstract class MockitoTest {
 }
